requirements.txt
python-telegram-bot
openai
requests
python-dotenv
gitpython
apscheduler
gTTS
SpeechRecognition
